Refactor: 로또 생성기능 구매개수 인자형 수정
- 구매 개수가 상당히 많을수도 있다는 가정하에 count를 int에서 long으로 수정.

+++ src/main/java/lotto/LottoGenerator.java
@@ -2,16 +2,16 @@ package lotto;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public enum LottoGenerator {
     INSTANCE;
-    private final int BOTTOM_RANGE=1;
-    private final int TOP_RANGE=45;
+    private final int BOTTOM_RANGE = 1;
+    private final int TOP_RANGE = 45;
 
-    public List<Lotto> generate(int count){
-        return IntStream.range(0,count)
-                .mapToObj(generateCount->new Lotto(Randoms.pickUniqueNumbersInRange(BOTTOM_RANGE,TOP_RANGE,6)))
+    public List<Lotto> generate(long count) {
+        return LongStream.range(0, count)
+                .mapToObj(generateCount -> new Lotto(Randoms.pickUniqueNumbersInRange(BOTTOM_RANGE, TOP_RANGE, 6)))
                 .toList();
     }
 }
